:zap: update demo examples
