+++ examples/retrieval_qa_w_sources/requirements.txt
@@ -1,6 +1,7 @@
 langchain
 openai
+tiktoken
+faiss-cpu
+gradio
 fastapi
 uvicorn[standard]
-faiss-cpu
-tiktoken
